// apps/recommendation-items-service/src/__generated__/resolvers-types.d.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
    ID: {
        input: string;
        output: string;
    };
    String: {
        input: string;
        output: string;
    };
    Boolean: {
        input: boolean;
        output: boolean;
    };
    Int: {
        input: number;
        output: number;
    };
    Float: {
        input: number;
        output: number;
    };
    DateTime: {
        input: any;
        output: any;
    };
    _FieldSet: {
        input: any;
        output: any;
    };
};
export type CreateRecommendationItemInput = {
    category: Scalars['String']['input'];
    contraindications: Array<Scalars['String']['input']>;
    description: Scalars['String']['input'];
    evidenceLevel: EvidenceLevel;
    guidelines: Array<Scalars['String']['input']>;
    instructions?: InputMaybe<Scalars['String']['input']>;
    sideEffects: Array<Scalars['String']['input']>;
    studyReferences: Array<Scalars['String']['input']>;
    title: Scalars['String']['input'];
    type: RecommendationItemType;
};
export declare enum EvidenceLevel {
    Consensus = "CONSENSUS",
    ExpertOpinion = "EXPERT_OPINION",
    LevelI = "LEVEL_I",
    LevelIi = "LEVEL_II",
    LevelIii = "LEVEL_III",
    LevelIv = "LEVEL_IV",
    LevelV = "LEVEL_V"
}
// ...
export declare enum RecommendationItemType {
    Education = "EDUCATION",
    FollowUp = "FOLLOW_UP",
    Imaging = "IMAGING",
    LabTest = "LAB_TEST",
    Lifestyle = "LIFESTYLE",
    Medication = "MEDICATION",
    Procedure = "PROCEDURE",
    Screening = "SCREENING",
    Therapy = "THERAPY",
    Vaccination = "VACCINATION"
}
export type UpdateRecommendationItemInput = {
    category?: InputMaybe<Scalars['String']['input']>;
    contraindications?: InputMaybe<Array<Scalars['String']['input']>>;
    description?: InputMaybe<Scalars['String']['input']>;
    evidenceLevel?: InputMaybe<EvidenceLevel>;
    guidelines?: InputMaybe<Array<Scalars['String']['input']>>;
    instructions?: InputMaybe<Scalars['String']['input']>;
    isActive?: InputMaybe<Scalars['Boolean']['input']>;
    sideEffects?: InputMaybe<Array<Scalars['String']['input']>>;
    studyReferences?: InputMaybe<Array<Scalars['String']['input']>>;
    title?: InputMaybe<Scalars['String']['input']>;
    type?: InputMaybe<RecommendationItemType>;
};